// src/app/services/models/project-to-send.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProjectToSend {
  date?: string;
  description?: string;
  domaine?: 'Web' | 'CyberSecurity' | 'Ai_Models';
  isOpen?: boolean;
  isVisible?: boolean;
  name?: string;
  professorId?: number;
  studentId?: number;
  report?: File;
  video?: File;
  images?: Array<File>;

}